organised files in the codebase based on components
