+++ src/app/components/pages/IsaacCompetition/Buttons/CompetitionButton.tsx
